skill-factory/generated_skills/global-market-liquidity-analyst/scripts/requirements.txt
pandas>=1.5.0
numpy>=1.23.0
openpyxl>=3.0.0
python-docx>=1.1.0

